#2  Added CDR
Added cooldown reduction -
We check if time a key is a pressed.
If he was pressed 10 times we block him 10 times:
The block is that we check in the start of the function for the counter, and if a state is true then he blocks from checking the key pressing.
After 10 presses he allows to shot